Ticket #223 and #212: Add the ability to print Ext grid and fix backgroundColor overriding the content
git-svn-id: http://www.mapfish.org/svn/mapfish/trunk/MapFish/server/java/print@1026 25cca10d-0837-0410-9e4f-92b931fda1bc

+++ print-lib/src/main/java/org/mapfish/print/scalebar/Label.java
@@ -23,11 +23,11 @@ public class Label {
      */
     public final float height;
 
-    public Label(float paperOffset, String label, BaseFont font, float fontSize, boolean rotated) {
+    public Label(float paperOffset, String label, BaseFont font, double fontSize, boolean rotated) {
         this.paperOffset = paperOffset;
         this.label = label;
-        final float textWidth = font.getWidthPoint(label, fontSize);
-        final float textHeight = font.getAscentPoint(label, fontSize) - font.getDescentPoint(label, fontSize);
+        final float textWidth = font.getWidthPoint(label, (float) fontSize);
+        final float textHeight = font.getAscentPoint(label, (float) fontSize) - font.getDescentPoint(label, (float) fontSize);
         if (rotated) {
             this.height = textWidth;
             this.width = textHeight;
